+++ CHANGELOG.md
@@ -17,6 +17,7 @@ All notable changes to this project will be documented in this file.
 - [#2259] Optional resampling sound to 44kHz on playback and on export
 - Set class to character mapping tool (via SymbolClass) context menu on characters
 - [#2189] Search bar in replace character (+ replace references) window
+- [#2011], [#2215] Option to ignore frame background color when exporting (make transparent)
 
 ### Fixed
 - Debugger - getting children of top level variables
@@ -3410,6 +3411,8 @@ Major version of SWF to XML export changed to 2.
 [#2185]: https://www.free-decompiler.com/flash/issues/2185
 [#2259]: https://www.free-decompiler.com/flash/issues/2259
 [#2189]: https://www.free-decompiler.com/flash/issues/2189
+[#2011]: https://www.free-decompiler.com/flash/issues/2011
+[#2215]: https://www.free-decompiler.com/flash/issues/2215
 [#2149]: https://www.free-decompiler.com/flash/issues/2149
 [#2172]: https://www.free-decompiler.com/flash/issues/2172
 [#2174]: https://www.free-decompiler.com/flash/issues/2174